services/orchestrator/requirements.txt
# Core dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# HTTP client
httpx==0.28.1
requests==2.31.0
pandas>=2.0.0
pyarrow>=12.0.0
fastparquet>=2023.0.0

# OpenAI integration
openai>=1.0.0

# Environment and logging
python-dotenv==1.0.1
structlog==25.4.0

# System monitoring
psutil==6.1.0

# Authentication and security
python-multipart==0.0.6

# Metrics and monitoring
prometheus-client==0.19.0
redis==5.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code quality
ruff==0.8.5
mypy==1.14.1

# API generation
openapi-generator-cli==7.10.0

# Smart Home (Lexi integration)
python-kasa>=0.5.0

# Web Automation (Lexi integration)
playwright>=1.40.0